Reorganize component folder structure
