+++ components/layout/FactCard.tsx
