+++ components/layout/GetStartedStep.tsx
