+++ components/layout/Header.tsx
@@ -20,8 +20,7 @@ import React, { useState } from 'react';
 
 import navItems from '../../content/navItems.ts';
 // import GivingTuesdayBanner from '../GivingTuesdayBanner.tsx';
-import NavigationLogo from '../NavigationLogo.tsx';
-import SkipLink from '../SkipLink.tsx';
+import NavigationLogo from './NavigationLogo.tsx';
 
 export default function Header() {
   const theme = useTheme();
@@ -67,7 +66,6 @@ export default function Header() {
       <AppBar className="nav-desktop" color="primary" elevation={0} component="nav" position="sticky">
         <Container maxWidth="xl" sx={{ p: 3 }}>
           <Toolbar disableGutters sx={{ height: { xs: 64 } }}>
-            <SkipLink color="primary" variant="contained" />
             {/* Boxes as containers for handling layout among siblings */}
             <Box
               sx={{
