+++ components/layout/NavigationLogo.tsx
