+++ components/layout/PaperSection.tsx
