+++ components/layout/ResearchBanner.tsx
